chore: ts 초기설정
