+++ package.json
@@ -1,0 +1,23 @@
+{
+	"name": "todolist",
+	"version": "1.0.0",
+	"description": "todolist",
+	"main": "index.js",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/clapho/TodoList.git"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"bugs": {
+		"url": "https://github.com/clapho/TodoList/issues"
+	},
+	"homepage": "https://github.com/clapho/TodoList#readme",
+	"devDependencies": {
+		"typescript": "^5.0.4"
+	}
+}
