+++ src/js/TodoList.ts
@@ -1,4 +1,4 @@
-import Validate from './Validation.js';
+import Validate from './Validation';
 
 /**
  *
