+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"compilerOptions": {
+		"allowJs": true,
+		"target": "ES5",
+		"outDir": "./built",
+		"moduleResolution": "node",
+		"strict": true
+	},
+	"include": ["./src/**/*"]
+}
